Native lib - reorganize components in folders for convenience + bump version to 5.0.0
