+++ lib/components/HighlighterOverlay/HighlighterOverlayView.web.tsx
