+++ lib/components/Keyboard/KeyboardRegistry/EventEmitterManager/__tests__/EventEmitterManager.spec.js
@@ -1,4 +1,4 @@
-const EventEmitterManager = require('../EventEmitterManager').default;
+const EventEmitterManager = require('..').default;
 
 let EventEmitter;
