+++ lib/components/Keyboard/KeyboardRegistry/__tests__/KeyboardRegistry.spec.js
@@ -1,6 +1,6 @@
 import {AppRegistry, View} from 'react-native';
 import React from 'react';
-import KeyboardRegistry from '../KeyboardRegistry';
+import KeyboardRegistry from '..';
 
 describe('KeyboardRegistry - components', () => {
   const mockComponent = 'test_component';
