+++ lib/components/Keyboard/KeyboardRegistry/keyboardRegistry.api.json
