+++ lib/components/Keyboard/KeyboardTrackingView/KeyboardTrackingView.android.tsx
