+++ lib/components/Keyboard/index.ts
@@ -1,15 +1,15 @@
-import KeyboardTrackingView, {KeyboardTrackingViewProps} from './KeyboardTracking/KeyboardTrackingView';
-import KeyboardAwareInsetsView from './KeyboardTracking/KeyboardAwareInsetsView';
-import KeyboardRegistry from './KeyboardInput/KeyboardRegistry';
-import KeyboardAccessoryView, {KeyboardAccessoryViewProps} from './KeyboardInput/KeyboardAccessoryView';
-import KeyboardUtils from './KeyboardInput/utils/KeyboardUtils';
+import KeyboardTrackingView, {KeyboardTrackingViewProps} from './KeyboardTrackingView';
+import KeyboardAwareInsetsView from './KeyboardAwareInsetsView';
+import KeyboardAccessoryView, {KeyboardAccessoryViewProps} from './KeyboardAccessoryView';
+import KeyboardRegistry from './KeyboardRegistry';
+import KeyboardUtils from './KeyboardUtils';
 
 export {KeyboardTrackingViewProps, KeyboardAccessoryViewProps};
 
 export default {
   KeyboardTrackingView,
   KeyboardAwareInsetsView,
-  KeyboardRegistry,
   KeyboardAccessoryView,
+  KeyboardRegistry,
   KeyboardUtils
 };
